refactor(cmd): decompose serve_find_smells.go via /evolve [evolve-simplify-find-smells]
Pure-move decomposition of cmd/serve_find_smells.go (1328 → 230 LOC orchestrator) via /rosary:evolve --simplify --focus. 5-file decomposition:
- serve_find_smells.go (230) — handler orchestrator + rule dispatch + discovery payload
- smell_rules.go (696) — SmellRule type + smellRegistry + 9 built-in rules
- smell_rules_external.go (51) — init() + env-driven external rule loading
- smell_refs_views.go (244) — canonical views (v_defs/v_refs) + binding readthrough (ADR-0013)
- smell_findings.go (129) — rule execution + filtering + snippet population

30 // coverage:ignore annotations in the residual orchestrator file — skeptic-agent verified 6/6 random spot-checks against base coverage; every annotated line is pre-existing-uncovered (count=0 on base branch). NOT new debt.

ADR-0017 I3 invariant preserved: untested_function rule (ID, Languages, Requires, ScopeColumn, Description, full Query body) is byte-for-byte identical to base. The test gold (cmd/serve_find_smells_test.go, 2996 LOC, 9 rules) passes unchanged. cmd/serve_find_smells_views_test.go (v_defs/v_refs shape) passes unchanged. cmd/serve_find_smells_bench_test.go passes.

Verification (per evaluator):
- gofumpt + go vet + golangci-lint: 0 issues
- go test -tags boltdb -race ./cmd/...: PASS (130.9s)
- coverage-gate (--rename-threshold 50, --cached): exit 0
- task check: exit 0
- Coverage 60.9% (baseline preserved)
- Pre-existing flake (TestFindSmellsCLI_DefaultMinMetricApplied under -race -count=2 due to cobra global state) reproduces on base — not introduced.

Annotation surface tracked in mache-89b5dd (P1 epic, post-decomposition reduction campaign).

Closes ephemeral beads filed by scoping-agent.
